+++ cmd/tendermint/commands/run_node.go
@@ -10,9 +10,9 @@ import (
 	"github.com/spf13/cobra"
 
 	cfg "github.com/celestiaorg/celestia-core/config"
-	"github.com/celestiaorg/celestia-core/ipfs"
 	tmos "github.com/celestiaorg/celestia-core/libs/os"
 	nm "github.com/celestiaorg/celestia-core/node"
+	"github.com/celestiaorg/celestia-core/pkg/da/ipfs"
 )
 
 var (
