+++ consensus/reactor_test.go
@@ -23,7 +23,6 @@ import (
 	cstypes "github.com/celestiaorg/celestia-core/consensus/types"
 	cryptoenc "github.com/celestiaorg/celestia-core/crypto/encoding"
 	"github.com/celestiaorg/celestia-core/crypto/tmhash"
-	"github.com/celestiaorg/celestia-core/ipfs"
 	"github.com/celestiaorg/celestia-core/libs/bits"
 	"github.com/celestiaorg/celestia-core/libs/bytes"
 	"github.com/celestiaorg/celestia-core/libs/db/memdb"
@@ -32,6 +31,7 @@ import (
 	mempl "github.com/celestiaorg/celestia-core/mempool"
 	"github.com/celestiaorg/celestia-core/p2p"
 	p2pmock "github.com/celestiaorg/celestia-core/p2p/mock"
+	"github.com/celestiaorg/celestia-core/pkg/da/ipfs"
 	tmproto "github.com/celestiaorg/celestia-core/proto/tendermint/types"
 	sm "github.com/celestiaorg/celestia-core/state"
 	statemocks "github.com/celestiaorg/celestia-core/state/mocks"
