+++ node/node.go
@@ -24,7 +24,6 @@ import (
 	cs "github.com/celestiaorg/celestia-core/consensus"
 	"github.com/celestiaorg/celestia-core/crypto"
 	"github.com/celestiaorg/celestia-core/evidence"
-	"github.com/celestiaorg/celestia-core/ipfs"
 	dbm "github.com/celestiaorg/celestia-core/libs/db"
 	"github.com/celestiaorg/celestia-core/libs/db/badgerdb"
 	tmjson "github.com/celestiaorg/celestia-core/libs/json"
@@ -35,6 +34,7 @@ import (
 	mempl "github.com/celestiaorg/celestia-core/mempool"
 	"github.com/celestiaorg/celestia-core/p2p"
 	"github.com/celestiaorg/celestia-core/p2p/pex"
+	"github.com/celestiaorg/celestia-core/pkg/da/ipfs"
 	"github.com/celestiaorg/celestia-core/privval"
 	"github.com/celestiaorg/celestia-core/proxy"
 	rpccore "github.com/celestiaorg/celestia-core/rpc/core"
