+++ pkg/da/ipld/validate.go
