+++ types/block.go
@@ -21,7 +21,7 @@ import (
 	tmmath "github.com/celestiaorg/celestia-core/libs/math"
 	"github.com/celestiaorg/celestia-core/libs/protoio"
 	tmsync "github.com/celestiaorg/celestia-core/libs/sync"
-	"github.com/celestiaorg/celestia-core/p2p/ipld/wrapper"
+	"github.com/celestiaorg/celestia-core/pkg/da/ipld/wrapper"
 	tmproto "github.com/celestiaorg/celestia-core/proto/tendermint/types"
 	tmversion "github.com/celestiaorg/celestia-core/proto/tendermint/version"
 	"github.com/celestiaorg/celestia-core/types/consts"
